app(refactor): Align cancel run button with start run
