+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/ActionButton/index.tsx
@@ -133,6 +133,7 @@ export function ActionButton(props: ActionButtonProps): JSX.Element {
         }
         onClick={handleButtonClick}
         id="ProtocolRunHeader_runControlButton"
+        borderRadius="200px"
         {...targetProps}
       >
         {buttonIconName != null ? (
