+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/RunHeaderSectionUpper.tsx
@@ -2,17 +2,21 @@ import { useTranslation } from 'react-i18next'
 import { useNavigate } from 'react-router-dom'
 import { css } from 'styled-components'
 
-import { RUN_STATUS_IDLE } from '@opentrons/api-client'
+import { RUN_STATUS_IDLE, RUN_STATUS_RUNNING } from '@opentrons/api-client'
 import {
   Box,
+  COLORS,
   DISPLAY_GRID,
   Flex,
   JUSTIFY_FLEX_END,
+  PrimaryButton,
   SecondaryButton,
   SPACING,
+  StyledText,
 } from '@opentrons/components'
 
 import { RunTimer } from '/app/molecules/RunTimer'
+import { useRunControls } from '/app/organisms/RunTimeControl/hooks'
 import { useFeatureFlag } from '/app/redux/config'
 import {
   useProtocolDetailsForRun,
@@ -21,6 +25,7 @@ import {
 } from '/app/resources/runs'
 
 import { DisplayRunStatus } from '../DisplayRunStatus'
+import { isCancellableStatus } from '../utils'
 import { ActionButton } from './ActionButton'
 import { LabeledValue } from './LabeledValue'
 
@@ -30,11 +35,11 @@ import type { RunHeaderContentProps } from '.'
 export function RunHeaderSectionUpper(
   props: RunHeaderContentProps
 ): JSX.Element {
-  const { runId, runStatus, robotName } = props
-
+  const { runId, runStatus, robotName, runHeaderModalContainerUtils } = props
   const { t } = useTranslation('run_details')
   const enableProtocolTimeline = useFeatureFlag('protocolTimeline')
   const navigate = useNavigate()
+  const { pause } = useRunControls(runId)
 
   const createdAtTimestamp = useRunCreatedAtTimestamp(runId)
   const { startedAt, stoppedAt, completedAt } = useRunTimestamps(runId)
@@ -45,6 +50,12 @@ export function RunHeaderSectionUpper(
     const encodedTimestamp = encodeURIComponent(createdAtTimestamp)
     const targetPath = `/devices/${robotName}/${runId}/${encodedTimestamp}/${protocolKey}/visualization`
     navigate(targetPath)
+  }
+  const handleCancelRunClick = (): void => {
+    if (runStatus === RUN_STATUS_RUNNING) {
+      pause()
+    }
+    runHeaderModalContainerUtils.confirmCancelModalUtils.toggleModal()
   }
 
   return (
@@ -74,7 +85,19 @@ export function RunHeaderSectionUpper(
             {t('visualize')}
           </SecondaryButton>
         ) : null}
-        <ActionButton {...props} />
+        <Flex gridGap={SPACING.spacing8}>
+          {isCancellableStatus(runStatus) && (
+            <PrimaryButton
+              borderRadius="200px"
+              onClick={handleCancelRunClick}
+              id="RunHeader_cancelRunButton"
+              style={{ backgroundColor: COLORS.red50 }}
+            >
+              <StyledText as="pSemiBold">{t('cancel_run')}</StyledText>
+            </PrimaryButton>
+          )}
+          <ActionButton {...props}></ActionButton>
+        </Flex>
       </Flex>
     </Box>
   )
